GCC & Clang: enable a lot of compiler warnings and fixing them all (#465)

+++ sdk/core/core/test/cmocka/test_http_request.c
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "az_json_string_private.h"
+#include "az_test_definitions.h"
 #include <az_http.h>
 #include <az_http_internal.h>
 #include <az_http_private.h>
@@ -83,14 +84,14 @@ void test_http_request(void** state)
     TEST_EXPECT_SUCCESS(az_http_request_append_header(
         &hrb, hrb_header_authorization_name, hrb_header_authorization_token1));
 
-    assert_true(az_span_length(hrb._internal.headers) / sizeof(az_pair) == 2);
+    assert_true(az_span_length(hrb._internal.headers) / (int32_t)sizeof(az_pair) == 2);
     assert_true(hrb._internal.retry_headers_start_byte_offset == sizeof(az_pair));
 
     az_pair expected_headers1[2] = {
       { .key = hrb_header_content_type_name, .value = hrb_header_content_type_token },
       { .key = hrb_header_authorization_name, .value = hrb_header_authorization_token1 },
     };
-    for (uint16_t i = 0; i < az_span_length(hrb._internal.headers) / sizeof(az_pair); ++i)
+    for (uint16_t i = 0; i < az_span_length(hrb._internal.headers) / (int32_t)sizeof(az_pair); ++i)
     {
       az_pair header = { 0 };
       TEST_EXPECT_SUCCESS(az_http_request_get_header(&hrb, i, &header));
@@ -104,14 +105,14 @@ void test_http_request(void** state)
 
     TEST_EXPECT_SUCCESS(az_http_request_append_header(
         &hrb, hrb_header_authorization_name, hrb_header_authorization_token2));
-    assert_true(az_span_length(hrb._internal.headers) / sizeof(az_pair) == 2);
+    assert_true(az_span_length(hrb._internal.headers) / (int32_t)sizeof(az_pair) == 2);
     assert_true(hrb._internal.retry_headers_start_byte_offset == sizeof(az_pair));
 
     az_pair expected_headers2[2] = {
       { .key = hrb_header_content_type_name, .value = hrb_header_content_type_token },
       { .key = hrb_header_authorization_name, .value = hrb_header_authorization_token2 },
     };
-    for (uint16_t i = 0; i < az_span_length(hrb._internal.headers) / sizeof(az_pair); ++i)
+    for (uint16_t i = 0; i < az_span_length(hrb._internal.headers) / (int32_t)sizeof(az_pair); ++i)
     {
       az_pair header = { 0 };
       TEST_EXPECT_SUCCESS(az_http_request_get_header(&hrb, i, &header));
